Remove support for merging
This feature needs more time or proper CLS support. It would break whenever non-equal geometry would have to be merged, when merges are nested. and in general it would just represent faking a graph algorithm on a structure that is a tree. CLS would need to support equality or such a merge indicator for this to properly work.